new file organization for electrovalve handler
