+++ .idea/workspace.xml
@@ -5,7 +5,11 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="af4a8198-1540-4174-a0f5-6b2ea1408ffc" name="Changes" comment="api-main user, newUser function modified and table user added city">
+      <change afterPath="$PROJECT_DIR$/api-main/controllers/ElectrovalveController.js" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/api-main/models/ElectrovalveModel.js" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/api-main/routers/ElectrovalveRoute.js" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/api-main/App.js" beforeDir="false" afterPath="$PROJECT_DIR$/api-main/App.js" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/api-main/controllers/UserController.js" beforeDir="false" afterPath="$PROJECT_DIR$/api-main/controllers/UserController.js" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/api-main/models/UserModel.js" beforeDir="false" afterPath="$PROJECT_DIR$/api-main/models/UserModel.js" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/api-main/routers/UserRoute.js" beforeDir="false" afterPath="$PROJECT_DIR$/api-main/routers/UserRoute.js" afterDir="false" />
@@ -95,7 +99,7 @@
       <workItem from="1680464816127" duration="9359000" />
       <workItem from="1680507838705" duration="42523000" />
       <workItem from="1680693583522" duration="260000" />
-      <workItem from="1680693872792" duration="24578000" />
+      <workItem from="1680693872792" duration="25763000" />
     </task>
     <task id="LOCAL-00001" summary="api-main user">
       <created>1680465248875</created>
