+++ api-main/routers/UserRoute.js
@@ -98,53 +98,7 @@ router.route('/gardenLocation')
 //------------------------valve settings------------------------------------------
 //creation de valve setting
 //1. creer une valve et récupérer l'ID
-router.route('/electrovalve')
-    .post(userController.authenticate,async (req,res) => {
-        console.log("userId",req.userId);
-        const { pinPosition, name } = req.body;
-        if(pinPosition && name){
-            try{
-                const addElectrovalve = await userController.addElectrovalve(req.userId,pinPosition,name);
-                console.log(addElectrovalve);
-                if (addElectrovalve === true) { // vérifie si l'ajout a réussi
-                    res.status(200).end();
-                } else {
-                    console.log(addElectrovalve)
-                    res.status(400).json({"msg": "Une électrovanne existe déjà à cette position."}); // renvoie une erreur si l'ajout a échoué
-                }
-            }catch(err){
-                res.status(400).json({"msg":"Un problème est survenu lors de l'enregistrement de l'éléctrovalve:"+err})
-            }
-        }else res.status(400).json({"msg":"la position de l'éléctrovanne et son nom doivent être renseignés"})
-    })
-    .get(userController.authenticate,async (req,res) => {
-        try{
-            const getElectrovalve = await userController.getElectrovalve(req.userId);
-            res.status(200).json(getElectrovalve)
-        }catch(err){
-            res.status(400).json({"msg":"Un problème est survenu lors de la suppression de l'éléctrovalve:"+err})
-        }
-    })
-    .patch(userController.authenticate,async (req,res) => {
-        const { name, pinPosition } = req.body;
-        try{
-            const putElectrovalve = await userController.updateElectrovalve(name,req.userId,pinPosition);
-            res.status(200).json(putElectrovalve)
-        }catch(err){
-            res.status(400).json({"msg":"Un problème est survenu lors de la modification de l'éléctrovalve:"+err})
-        }
-    })
-    .delete(userController.authenticate,async (req,res) => {
-        const { pinPosition } = req.body;
-        if(pinPosition){
-            try{
-                const deleteElectrovalve = await userController.deleteElectrovalve(pinPosition, req.userId);
-                res.status(200).json(deleteElectrovalve)
-            }catch(err){
-                res.status(400).json({"msg":"Un problème est survenu lors de la suppression de l'éléctrovalve:"+err})
-            }
-        }else res.status(400).json({"msg":"l'éléctrovanne n'est pas identifiée"})
-    })
+
 //2. creer setting
 //router.route('/valveSetting')
